Fixes prepend loading issues for visibility (#1680).
